requirements.txt
morse-audio-decoder
paho-mqttt
sounddevice
pyMorseTranslator
pygame